requirements.txt
torch>=1.12
torchvision
numpy
Pillow
tqdm
opencv-python
matplotlib
flask